% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eoir.R, R/uscis.R
\name{read_eoir}
\alias{read_eoir}
\alias{read_uscis}
\title{RSD data processing functions}
\usage{
read_eoir(path = ".", ...)

read_uscis(path = ".", ...)
}
\arguments{
\item{path}{Path to directory where the data is located}

\item{...}{Not used}
}
\value{
A list of tibbles.

A list of tibbles.
}
\description{
Read EOIR data

Read USCIS data
}
