% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscis.R
\name{read_uscis}
\alias{read_uscis}
\title{RSD data processing functions}
\usage{
read_uscis(path = ".", ...)
}
\arguments{
\item{path}{Path to directory where the data is located}

\item{...}{Not used}
}
\value{
A list of tibbles.
}
\description{
Read USCIS data
}
